move adapter test registration out of t.action() into module scope
convex-test 0.0.45+ tracks test context via AsyncLocalStorage instead of a
global. Registering the adapter compliance suite from inside an action
(t.action(api.adapterTest.runTests, {})) leaves the ctx unreachable by the
time vitest runs the deferred test() callbacks the suite created.

Run the suite at module scope of adapter.test.ts and pass convexAdapter a
wrapped ctx that delegates each runQuery / runMutation to t.query /
t.mutation, so each adapter call enters its own ALS frame.

Tested on convex-test 0.0.41 and 0.0.50: 168 pass / 135 skip on both.
